make view len 16 bit

+++ ednsview.go
@@ -17,7 +17,7 @@ func (vo *ViewOpt) Rend(render *MsgRender) {
 	render.WriteUint16(EDNS_VIEW)
 	viewLen := len(vo.view)
 	render.WriteUint16(uint16(1 + viewLen))
-	render.WriteUint8(uint8(len(vo.view)))
+	render.WriteUint16(uint16(len(vo.view)))
 	render.WriteData([]byte(vo.view))
 }
 
@@ -27,7 +27,7 @@ func (vo *ViewOpt) String() string {
 
 //read from OPTION-LENGTH
 func viewOptFromWire(buffer *util.InputBuffer) (Option, error) {
-	l, err := buffer.ReadUint8()
+	l, err := buffer.ReadUint16()
 	if err != nil {
 		return nil, err
 	}
